Update github-status extension
