+++ extensions/github-status/package.json
@@ -5,7 +5,14 @@
   "description": "Quickly check why GitHub is down today",
   "icon": "command-icon.png",
   "author": "FezVrasta",
+  "contributors": [
+    "j3lte"
+  ],
   "license": "MIT",
+  "platforms": [
+    "macOS",
+    "Windows"
+  ],
   "keywords": [
     "Productivity"
   ],
@@ -19,19 +26,16 @@
     }
   ],
   "dependencies": {
-    "@raycast/api": "^1.31.0",
-    "node-fetch": "^3.2.3",
-    "react-use": "^17.3.2"
+    "@raycast/api": "^1.104.8",
+    "@raycast/utils": "^2.2.3"
   },
   "devDependencies": {
-    "@types/node": "~16.10.0",
-    "@types/react": "^17.0.28",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "prettier": "^2.5.1",
-    "typescript": "^4.4.3"
+    "@raycast/eslint-config": "^2.1.1",
+    "@types/node": "22.13.10",
+    "@types/react": "19.0.10",
+    "eslint": "^10.0.2",
+    "prettier": "^3.8.1",
+    "typescript": "^5.9.3"
   },
   "scripts": {
     "build": "ray build -e dist",
